Added OpenOffice/Ubuntu file bug (#38)
The earlier (and actual) bug report regards the __file__ utility, but I like the OpenOffice report as an example of how such a bug manifests itself to the end user. Most engineers when hearing the OpenOffice report would think "That can't possibly be right!" but how do you get to the root cause?